% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_bipoc}
\alias{add_bipoc}
\title{Adds BIPOC}
\usage{
add_bipoc(df)
}
\arguments{
\item{df}{Data frame. Requires \code{castaway_id}.}
}
\value{
Data frame with BIPOC added.
}
\description{
Adds a BIPOC to the data frame. If any African American, Asian American,
Latin American, or Native American is \code{TRUE} then BIPOC is \code{TRUE}.
}
\examples{

library(survivoR)
library(dplyr)

get_cast("US47") |>
  add_bipoc()

}
