% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_castaway}
\alias{add_castaway}
\title{Add castaway}
\usage{
add_castaway(df)
}
\arguments{
\item{df}{Data frame. Requires \code{castaway_id}.}
}
\value{
Data frame with \code{castaway}.
}
\description{
Adds \code{castaway} to a data frame. Input data frame must have \code{castaway_id}.
}
\examples{

library(survivoR)
library(dplyr)

df_no_castaway <- confessionals |>
  filter_us(47) |>
  group_by(castaway_id) |>
  summarise(n = sum(confessional_count))

df_no_castaway |>
  add_castaway()

}
