% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_demogs}
\alias{add_demogs}
\title{Add demographics}
\usage{
add_demogs(df)
}
\arguments{
\item{df}{Data frame. Requires \code{castaway_id}.}
}
\value{
Data frame with castaway added to it.
}
\description{
Add demographics that includes age, gender, race/ethnicity, and lgbtqia+
status to a data frame with \code{castaway_id}.
}
\examples{

library(survivoR)
library(dplyr)

get_cast("US47") |>
  add_demogs()

}
