% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_full_name}
\alias{add_full_name}
\title{Add full name}
\usage{
add_full_name(df)
}
\arguments{
\item{df}{Data frame. Requires \code{castaway_id}.}
}
\value{
Data frame with full name.
}
\description{
Adds full name to the data frame. Useful for plotting and making tables.
}
\examples{

library(survivoR)
library(dplyr)

get_cast("US47") |>
  add_full_name()

}
