% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_gender}
\alias{add_gender}
\title{Add gender}
\usage{
add_gender(df)
}
\arguments{
\item{df}{Data frame. Requires \code{castaway_id}.}
}
\value{
Data frame with gender added to it.
}
\description{
Adds gender to a data frame
}
\examples{

library(survivoR)
library(dplyr)

get_cast("US47") |>
  add_gender()

}
