% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{add_historical_data}
\alias{add_historical_data}
\title{Add Historical Data}
\usage{
add_historical_data(picks)
}
\arguments{
\item{picks}{Picks data from recent seasons}
}
\description{
Add Historical Data
}
