% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_jury}
\alias{add_jury}
\title{Add jury member}
\usage{
add_jury(df)
}
\arguments{
\item{df}{Data frame. Requires \code{version_season} and \code{castaway_id}.}
}
\value{
A data frame with a logical flag for the jury members
}
\description{
Adds a jury member flag to the data set.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  add_jury()

}
