% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_result}
\alias{add_result}
\title{Add result}
\usage{
add_result(df)
}
\arguments{
\item{df}{Data frame. Requires \code{castaway_id} and \code{version_season}.}
}
\value{
Data frame with result and place added.
}
\description{
Adds the result and place to the data frame.
}
\examples{

library(survivoR)
library(dplyr)

get_cast("US47") |>
  add_result()

}
