% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_tribe}
\alias{add_tribe}
\title{Add tribe}
\usage{
add_tribe(df, .tribe_status = "Original")
}
\arguments{
\item{df}{Data frame. Requires `version_season` and `castaway_id`,}

\item{.tribe_status}{Tribe status e.g. original, swapped, swapped_2, etc.}
}
\value{
Data frame with `tribe` added.
}
\description{
Adds tribe to a data frame for a specified stage of the game e.g. original,
swapped, swapped_2, etc.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  add_tribe()

}
