% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_tribe_colour}
\alias{add_tribe_colour}
\title{Add tribe colour}
\usage{
add_tribe_colour(df, .tribe_status = "Original")
}
\arguments{
\item{df}{Data frame. Requires \code{version_season} and \code{tribe}.}

\item{.tribe_status}{Tribe status e.g. original, swapped, swapped_2, etc.}
}
\value{
Data frame with \code{tribe_colour} added
}
\description{
Add tribe colour to the data frame. Useful for preparing the data for
plotting with ggplot2.
}
\examples{

library(survivoR)
library(dplyr)

get_cast("US47") |>
  add_tribe() |>
  add_tribe_colour()

}
