% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_winner}
\alias{add_winner}
\title{Add winner}
\usage{
add_winner(df)
}
\arguments{
\item{df}{Data frame. Requires \code{version_season} and \code{castaway_id}.}
}
\value{
A data frame with a logical flag for the winner
}
\description{
Adds a winner flag to the data set.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  add_winner()

}
