% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{advantage_timeline}
\alias{advantage_timeline}
\title{Advantage Timeline}
\format{
This data frame contains the following columns:
\describe{
\item{\code{version}}{Country code for the version of the show}
\item{\code{season}}{The season number}
\item{\code{version_season}}{Version season key}
\item{\code{episode}}{Episode}
\item{\code{sog_id}}{Stage of game ID}
\item{\code{castaway_id}}{Castaway ID}
\item{\code{castaway}}{Castaway}
\item{\code{advantage_id}}{The \code{advantage_id} key to map to the other \verb{advantage_*} tables}
\item{\code{final_n}}{FInal n players}
\item{\code{holding}}{1 if the player is holding the advantage at this stage of the game}
}
}
\usage{
advantage_timeline
}
\description{
A long format table to show who was holding which advantage/idol at what stage
of the game (\code{sog_id}). This table makes it simple to perform analysis on
advantages throughout the game.
}
\details{
\url{https://survivorstatsdb.com}
}
\keyword{datasets}
