% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{calculate_castaway_fields}
\alias{calculate_castaway_fields}
\title{Calculate Castaway Fields}
\usage{
calculate_castaway_fields(seasons_tbl)
}
\arguments{
\item{seasons_tbl}{Seasons table obtained from \code{process_seasons}}
}
\description{
Calculates the fields relevant to castaways for the pool, including their
placement, whether or not they were eliminated, and if they were the sole
survivor
}
