% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{calculate_participant_fields}
\alias{calculate_participant_fields}
\title{Calculate Payments & Rank}
\usage{
calculate_participant_fields(seasons_tbl)
}
\arguments{
\item{seasons_tbl}{Seasons table obtained from \code{process_seasons} and
processed by \code{calculate_castaway_fields}}
}
\description{
Calculates the payments and rank for the participants in the pool party
}
