% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{calculate_person_fields}
\alias{calculate_person_fields}
\title{Calculate Payments & Rank}
\usage{
calculate_person_fields(seasons_tbl, cost_per_day, sole_survivor_bonus)
}
\arguments{
\item{seasons_tbl}{Seasons table obtained from \code{process_seasons} and
processed by \code{calculate_castaway_fields}}

\item{cost_per_day}{Integer, number of dollars to multiply the day difference by}

\item{sole_survivor_bonus}{Integer, number of dollars to add to each persons cost
if one of the picks was the sole survivor}
}
\description{
Calculates the payments and rank for the people in the pool party
}
