% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_version}
\alias{check_version}
\title{Check package version}
\usage{
check_version()
}
\value{
A message
}
\description{
Compares the Github version to the system version currently loaded.
The user will be informed if more up to date data exists on Github
}
\examples{
check_version()
}
