% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{clean_all}
\alias{clean_all}
\title{Cleans all data sets}
\usage{
clean_all(
  folders = c("immunity", "jury-votes", "rewards", "season-cast", "vote-history"),
  source_loc = "dev/harvest"
)
}
\arguments{
\item{folders}{Folders of datasets to clean}

\item{source_loc}{Location of scripts to source}
}
\value{

}
\description{
Calls all scripts, cleans data for all seasons and binds data frames for package
}
\examples{
\dontrun{}
}
