% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_cast}
\alias{clean_cast}
\title{Clean cast data}
\usage{
clean_cast(df, season, season_name)
}
\arguments{
\item{df}{Data frame}

\item{season}{Season number}

\item{season_name}{Season name}
}
\value{

}
\description{
Clean cast data
}
\examples{
\dontrun{}
}
