% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_challenges}
\alias{clean_challenges}
\title{Clean challenges data}
\usage{
clean_challenges(df, season, season_name)
}
\arguments{
\item{df}{Data frame}

\item{season}{Season number}

\item{season_name}{Season name}
}
\value{

}
\description{
Clean challenges data
}
\examples{
\dontrun{}
}
