% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_viewers}
\alias{clean_viewers}
\title{Clean viewers data}
\usage{
clean_viewers(df, season, season_name)
}
\arguments{
\item{df}{Data frame}

\item{season}{Season number}

\item{season_name}{Season name}
}
\value{

}
\description{
Clean viewers data
}
\examples{
\dontrun{}
}
