% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_vote_matrix}
\alias{clean_vote_matrix}
\title{Clean vote history data}
\usage{
clean_vote_matrix(df, season, season_name, order)
}
\arguments{
\item{df}{Data frame}

\item{season}{Season number}

\item{season_name}{Season name}

\item{order}{Order of being voted out}
}
\value{

}
\description{
Clean vote history data
}
\examples{
\dontrun{}
}
