% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{cleanup_season_picks}
\alias{cleanup_season_picks}
\title{Cleanup Season Picks Dataset}
\usage{
cleanup_season_picks(season_picks)
}
\arguments{
\item{season_picks}{Season picks table after processing is applied}
}
\description{
Cleanup the season picks dataset after processing has been applied. This
includes removing unused fields and arranging columns and rows
}
