% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{create_season_participants}
\alias{create_season_participants}
\title{Create Participants Data Frame}
\usage{
create_season_participants(all_data)
}
\arguments{
\item{all_data}{Named list of tabs and their data from the survivor googlesheet}
}
\description{
Main method for creating the participants table with all transformations applied
}
