% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{create_season_picks}
\alias{create_season_picks}
\title{Create Season Picks Data Frame}
\usage{
create_season_picks(
  using = c("googlesheets", "configs"),
  config_path = "data-raw/configs",
  deauth_mode = FALSE
)
}
\arguments{
\item{using}{Either googlesheets or configs}

\item{config_path}{Path to the config files, defaults to the right path}

\item{deauth_mode}{Whether or not to authenticate with Google Sheets in deauth mode}
}
\description{
Main method for creating the season picks table with all transformations applied
}
