% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{.get_season_payments}
\alias{.get_season_payments}
\title{Get Participant Payments}
\usage{
.get_season_payments(picks, cost_per_day, sole_survivor_bonus)
}
\arguments{
\item{picks}{Picks dataset}

\item{cost_per_day}{Integer, number of dollars owed per day using the difference
in days compared to the person who's castaway lasted the longest}

\item{sole_survivor_bonus}{Integer, number of dollars to add to each persons
payment if the winners pick was the sole survivor}
}
\description{
Get Participant Payments
}
