% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{.picks_table__base_data}
\alias{.picks_table__base_data}
\title{Base Data for Season Picks Table}
\usage{
.picks_table__base_data(szn = default_season(), picks_only = FALSE)
}
\arguments{
\item{szn}{Season number. Defaults to \code{default_season()}}

\item{picks_only}{Logical, whether or not to only show castaways that have
been picked by a contestant}
}
\description{
Cleans up columns in season_picks table in preparation for formatting
}
