% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{.process_participant_name}
\alias{.process_participant_name}
\title{Parse Participant Name}
\usage{
.process_participant_name(picks)
}
\arguments{
\item{picks}{Picks dataset}
}
\description{
Parse Participant Name
}
