% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{.process_seasons__configs}
\alias{.process_seasons__configs}
\title{Process All Seasons Using Configs}
\usage{
.process_seasons__configs(config_path = "data-raw/configs")
}
\arguments{
\item{config_path}{Path to the config files, defaults to the right path}
}
\description{
Process each seasons config file with \code{process_season_config} and bind
all seasons together
}
