% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{.process_seasons__googlesheets}
\alias{.process_seasons__googlesheets}
\title{Process All Seasons Using Google Sheets}
\usage{
.process_seasons__googlesheets(deauth_mode = FALSE)
}
\arguments{
\item{deauth_mode}{Whether or not to authenticate with Google Sheets in deauth mode}
}
\description{
Process the data in the survivor Google Sheet
}
