% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{.winners_widgets__base_data}
\alias{.winners_widgets__base_data}
\title{Format Winners Data}
\usage{
.winners_widgets__base_data()
}
\description{
Gets table of winners and summarizes it for use by \code{format_winners_data}
}
