% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{episode_summary}
\alias{episode_summary}
\title{Episode summary}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{version}}{Country code for the version of the show}
  \item{\code{version_season}}{Version season key}
  \item{\code{episode}}{Episode number}
  \item{\code{episode_summary}}{summary of the episode}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
episode_summary
}
\description{
A dataset containing a summary of all US episodes seasons of Survivor
}
\keyword{datasets}
