% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{extract_vote_data}
\alias{extract_vote_data}
\title{Clean vote history data}
\usage{
extract_vote_data(df)
}
\arguments{
\item{df}{Data frame}
}
\value{

}
\description{
Clean vote history data
}
\examples{
\dontrun{}
}
