% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_au}
\alias{filter_au}
\title{Filter to AU seasons}
\usage{
filter_au(df, .season = NULL)
}
\arguments{
\item{df}{Data frame. Must include \code{version_season}.}

\item{.season}{Season or vector of seasons. If \code{NULL} if will filter to all US seasons.}
}
\value{
Data frame filtered to the specified US seasons
}
\description{
Filter a data set to a specified set of AU season or list of seasons. A
shorthand version of \code{filter_vs()} for the US seasons.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  filter_au(12)

}
