% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_jury}
\alias{filter_jury}
\title{Filter to jury}
\usage{
filter_jury(df)
}
\arguments{
\item{df}{Data frame. Requires \code{version_season} and \code{castaway_id}.}
}
\value{
A data frame filtered to the jury members
}
\description{
Filters a data set to the jury members of a given season.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  filter_jury()

}
