% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_new_era}
\alias{filter_new_era}
\title{Filter to the new era seasons}
\usage{
filter_new_era(df)
}
\arguments{
\item{df}{Data frame. Must include \code{version} and \code{season}.}
}
\value{
A data frame filtered to the New Era seasons.
}
\description{
Filters a data set to all New Era seasons.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  filter_new_era() |>
  distinct(version_season)

}
