% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_winner}
\alias{filter_winner}
\title{Filter to winners}
\usage{
filter_winner(df)
}
\arguments{
\item{df}{Data frame. Requires \code{version_season} and \code{castaway_id}.}
}
\value{
A data frame filtered to the winners
}
\description{
Filters a data set to the winners of a given season.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  filter_winner()

}
