% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{force_season_number}
\alias{force_season_number}
\title{Get Season Number}
\usage{
force_season_number(szn = default_season())
}
\arguments{
\item{szn}{Season number}
}
\description{
If all seasons are specified, then choose the default season
}
