% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_picks_table}
\alias{format_picks_table}
\title{Format Season Picks Table}
\usage{
format_picks_table(season = default_season())
}
\arguments{
\item{season}{Season number}
}
\description{
Selects, renames, and formats relevant columns for the app
}
