% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_cast}
\alias{get_cast}
\title{Get cast for a season}
\usage{
get_cast(.vs)
}
\arguments{
\item{.vs}{Version season. Can be a vector of \code{version_season} values.}
}
\value{
A data frame
}
\description{
For a given season (or seasons) the function will return a data frame of the cast.
}
\examples{
library(survivoR)

get_cast("US47")
}
