% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_castaways_remaining}
\alias{get_castaways_remaining}
\title{Get Number of Remaining Castaways}
\usage{
get_castaways_remaining(szn = default_season())
}
\arguments{
\item{szn}{Season number}
}
\description{
Get Number of Remaining Castaways
}
