% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{get_column_formats}
\alias{get_column_formats}
\title{Get Column Formats}
\usage{
get_column_formats(data)
}
\arguments{
\item{data}{Picks data that's ready for formatting}
}
\description{
Column formats for formattable table
}
