% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_confessional_timing}
\alias{get_confessional_timing}
\title{Confessional time}
\usage{
get_confessional_timing(paths, .vs, .episode)
}
\arguments{
\item{.vs}{Version season}

\item{.episode}{Episode}

\item{path}{Paths to the csv file containing all the time stamps from the Shiny app}
}
\value{
data frame
}
\description{
Takes the output of the times recorded from the Shiny app and aggregates to the final
confessional times and confessional counts. \code{confessional_time} is the total duration
in seconds for the episode. \code{confessional_count} is the number of confessionals
recorded to be at least 10 seconds apart.
}
\examples{
# After running app and recording confessionals, run...
# Example from a saved timing file

path <- system.file(package = "survivoR", "confessionalDash/2023-05-18 19h26m53s US4412.csv")
get_confessional_timing(path = path, .vs = "US44", .episode = 12)
}
