% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_pick_data}
\alias{get_pick_data}
\title{Filter the season_picks Table}
\usage{
get_pick_data(..., .szn = default_season())
}
\arguments{
\item{...}{Filter conditions, e.g. \code{castaway_name \%~\% 'rob'}}

\item{.szn}{Season number}
}
\value{
Data frame
}
\description{
Filters the season_picks table based on a set of condiions passed to \code{dplyr::filter}
}
