% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_pool_winner}
\alias{get_pool_winner}
\title{Get Season Pool Winner}
\usage{
get_pool_winner(szn = default_season())
}
\arguments{
\item{szn}{Season number. If all seasons specified, returns default season}
}
\value{
Named vector where name is the participant and value is their pick
}
\description{
Get Season Pool Winner
}
