% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_season_label}
\alias{get_season_label}
\title{Get Season Label}
\usage{
get_season_label(szn = default_season())
}
\arguments{
\item{szn}{Season number or \code{all_seasons_label()}}
}
\description{
Returns the display title for a season. If a meaningful name exists,
returns "Season X: Name". Otherwise returns "Season X".
}
