% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_season_participants}
\alias{get_season_participants}
\title{Get Season Picking Order}
\usage{
get_season_participants(szn = default_season())
}
\arguments{
\item{szn}{Integer, season number}
}
\description{
List participants and in what order they will submit their pick
}
