% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_season_picking_order}
\alias{get_season_picking_order}
\title{Get Season Picking Order}
\usage{
get_season_picking_order(szn = default_season(), pretty = TRUE)
}
\arguments{
\item{szn}{Integer, the season to find the picking order for}

\item{pretty}{Logical, whether to show the participants full name (TRUE, default)
or their ID (FALSE)}
}
\description{
Get the picking order for a season. This only applies when \code{szn} is an integer.
Notes:
\itemize{
\item When the previous season doesn't exist, all historical participants
are sorted randomly
\item Only participants in the \code{season_picks} table are included, so new participants
may need to be added manually
}
}
