% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_season_picks}
\alias{get_season_picks}
\title{List Season Castaways}
\usage{
get_season_picks(szn = default_season())
}
\arguments{
\item{szn}{Integer, season number}
}
\description{
List castaways for a given season
}
