% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{get_sole_survivor}
\alias{get_sole_survivor}
\title{Get Season Sole Survivor}
\usage{
get_sole_survivor(szn = default_season())
}
\arguments{
\item{szn}{Season number. If all seasons specified, returns default season}
}
\description{
Get Season Sole Survivor
}
