% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winners.R
\name{get_sole_survivors_table}
\alias{get_sole_survivors_table}
\title{Sole Survivors Table}
\usage{
get_sole_survivors_table()
}
\description{
Filter all seasons of the season_picks table to just the sole survivors
}
