% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{gs_auth}
\alias{gs_auth}
\title{Authenticate With Google Sheets}
\usage{
gs_auth(
  svc_acct_json_path = ".secrets/gcp-service-account.json",
  gargle_creds_env_name = "GCP_SERVICE_ACCOUNT_KEY_BASE64",
  cache = ".secrets",
  email = "elampsart@gmail.com",
  deauth_mode = FALSE
)
}
\arguments{
\item{svc_acct_json_path}{Path to service account json. Defaults to \code{.secrets/gcp-service-account.json}}

\item{gargle_creds_env_name}{Name of env var containing base64 encoded JSON service account}

\item{cache}{Path to OAuth cache directory}

\item{email}{Email for OAuth authentication}

\item{deauth_mode}{Logical: if TRUE, run \code{gs4_deauth()} and exit}
}
\description{
Tries multiple authentication methods in this order:
\enumerate{
\item Service account JSON file (local prod)
\item OAuth cache (local dev)
\item Base64-encoded service account from environment variable (CI/CD)
}
}
