% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{gs_get_all_data}
\alias{gs_get_all_data}
\title{Retrieve All Data From Google Sheets}
\usage{
gs_get_all_data(deauth_mode = FALSE)
}
\arguments{
\item{deauth_mode}{Whether or not to authenticate with Google Sheets in deauth mode}
}
\description{
Retrieve all relevant data from googlesheets
}
