% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{gs_get_srvivor_data}
\alias{gs_get_srvivor_data}
\title{Get Survivor Spreadsheet}
\usage{
gs_get_srvivor_data(tab_name)
}
\arguments{
\item{tab_name}{Name of the tab to extract}
}
\description{
Uses \code{\link[googlesheets4:gs4_get]{googlesheets4::gs4_get()}} to get nested data frame of sheets and tabs
}
