% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets.R
\name{gs_srvivor_spreadsheet}
\alias{gs_srvivor_spreadsheet}
\title{Get Survivor Spreadsheet}
\usage{
gs_srvivor_spreadsheet()
}
\description{
Uses \code{\link[googlesheets4:gs4_get]{googlesheets4::gs4_get()}} to get nested data frame of sheets and tabs
}
