% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historical_castaways}
\alias{historical_castaways}
\title{Historical Castaways}
\format{
A data frame with 785 rows and 7 variables:
\describe{
\item{\code{season}}{integer Season number}
\item{\code{castaway_id}}{character ID for the castaway}
\item{\code{castaway_name}}{character Full name of the castaway}
\item{\code{castaway_finish_day}}{integer The day the castaway was voted out}
\item{\code{castaway_eliminated}}{logical Whether or not survivor has been eliminated}
\item{\code{castaway_finish_placement}}{double Rank based on when castaway was voted out}
\item{\code{sole_survivor}}{logical Whether or not the castaway was the sole survivor}
}
}
\usage{
historical_castaways
}
\description{
Castaways from earlier seasons before we started our pool
}
\details{
Season configs processed into a table
}
\keyword{datasets}
