% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{import_non_us_data}
\alias{import_non_us_data}
\title{[EXPERIMENTAL] Import all Non-US seasons of SurvivoR}
\usage{
import_non_us_data()
}
\value{
Loads all versions of survivor into the package
}
\description{
[EXPERIMENTAL] Import all Non-US seasons of SurvivoR
}
\examples{
\dontrun{import_non_us_data()}
}
