% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{journeys}
\alias{journeys}
\title{Journeys}
\format{
This data frame contains the following columns:
\describe{
\item{\code{version}}{Country code for the version of the show}
\item{\code{season}}{The season number}
\item{\code{version_season}}{Version season key}
\item{\code{episode}}{Episode}
\item{\code{sog_id}}{Stage of game ID}
\item{\code{castaway_id}}{Castaway ID}
\item{\code{castaway}}{Castaway}
\item{\code{reward}}{The thing they won (or lost)}
\item{\code{lost_vote}}{Logical. If they lost their vote}
\item{\code{game_played}}{The game they played on the journey}
\item{\code{chose_to_play}}{If they chose to play or not}
\item{\code{event}}{The event that occured e.g. risked vote, lost vote}
}
}
\usage{
journeys
}
\description{
Details on who went on Journeys, what they won or if they lost their vote.
}
\details{
\url{https://survivorstatsdb.com}
}
\keyword{datasets}
