% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jury_votes}
\alias{jury_votes}
\title{Jury votes}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{episode}{Episode number of the reward challenge}
  \item{title}{Episode title}
  \item{immunity}{Winners of the immunity challenge}
}
}
\source{
\url{tba}
}
\usage{
jury_votes
}
\description{
A dataset containing details on the final jury votes to determine the winner for each season
}
\keyword{datasets}
