% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{last_voted_out}
\alias{last_voted_out}
\title{Last Castaway Voted Out}
\usage{
last_voted_out(szn = default_season())
}
\arguments{
\item{szn}{Season number}
}
\description{
Get the name of the last castaway voted out. If the season is over, returns
the two runners up. If the season just started, returns a string saying no
one has been voted out yet
}
