% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{launch_confessional_app}
\alias{launch_confessional_app}
\title{Launch Confessional App}
\usage{
launch_confessional_app(path = "confessional-timing", browser = TRUE)
}
\arguments{
\item{browser}{Open in browser instead of viewer. Logical.}
}
\value{
Shiny app
}
\description{
Opens a viewer or browser to time castaway confessionals
}
\examples{
# launch_confessional_app()
}
