% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{list_seasons}
\alias{list_seasons}
\title{List Seasons}
\usage{
list_seasons()
}
\description{
List the seasons available in this repository that have been set up
with a config file listing the survivors and picks
}
