% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_episode_transcripts}
\alias{load_episode_transcripts}
\title{Read episode transcripts}
\usage{
load_episode_transcripts()
}
\value{
A data frame of episode transcripts
}
\description{
Read the episode transcripts from Github. File is large and not explicitly part
of the package. Data is update by Matt Stiles.
}
\examples{
# Run
# load_episode_transcripts()
# to load all transcripts
}
