% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{picks_formatter}
\alias{picks_formatter}
\title{Format Picks Table Cells}
\usage{
picks_formatter(condition, color_yes = "green", color_no = "black")
}
\arguments{
\item{condition}{The condition to be evaluated to determine which color to use}

\item{color_yes}{If the condition evaluates to TRUE, use this color}

\item{color_no}{If the condition evaluates to FALSE, use this color}
}
\description{
Formats cells in the picks table given the row meets certain conditions. For
example, text is green if there is a winner
}
