% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{process_season_config}
\alias{process_season_config}
\title{Process Season Config}
\usage{
process_season_config(config, people)
}
\arguments{
\item{config}{Season config file parsed using \code{yaml::read_yaml}}

\item{people}{Config file of people in the pool parsed using \code{yaml::read_yaml}}
}
\description{
Process season config into a data frame
}
