% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{process_seasons}
\alias{process_seasons}
\title{Process Seasons Table}
\usage{
process_seasons(
  using = c("googlesheets", "configs"),
  config_path = "data-raw/configs",
  deauth_mode = FALSE
)
}
\arguments{
\item{using}{Either googlesheets or configs}

\item{config_path}{Path to the config files}

\item{deauth_mode}{Whether or not to authenticate with Google Sheets in deauth mode}
}
\description{
Creates the base season picks table using either the data in googlesheets
or the config files
}
