% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\name{read_webpages}
\alias{read_webpages}
\title{Read webpage}
\usage{
read_webpages(pages, out_loc = "dev/webpages/raw-html-data.rds", write = TRUE)
}
\arguments{
\item{pages}{List of webpages}

\item{out_loc}{Output location}

\item{write}{Write output? Logical}
}
\value{

}
\description{
Scrapes the Survivor wiki pages
}
\examples{
\dontrun{}
}
