% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_non_us_data}
\alias{remove_non_us_data}
\title{[EXPERIMENTAL] Remove Non-US seasons of SurvivoR}
\usage{
remove_non_us_data()
}
\value{
Removes the Non US data from the data frames
}
\description{
[EXPERIMENTAL] Remove Non-US seasons of SurvivoR
}
\examples{
\dontrun{remove_non_us_data()}
}
