% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{season_cast}
\alias{season_cast}
\title{Castaways}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{episode}{Episode number of the reward challenge}
  \item{title}{Episode title}
  \item{immunity}{Winners of the immunity challenge}
}
}
\source{
\url{tba}
}
\usage{
season_cast
}
\description{
A dataset containing details on the castaways for each season
}
\keyword{datasets}
