% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{season_castaways}
\alias{season_castaways}
\title{List Season Castaways}
\usage{
season_castaways(season)
}
\arguments{
\item{Season}{Integer, season number}
}
\description{
List castaways for a given season
}
