% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{season_has_started}
\alias{season_has_started}
\title{Check if Season Started}
\usage{
season_has_started(szn = default_season())
}
\arguments{
\item{szn}{Season number}
}
\description{
Check if Season Started
}
