% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\docType{data}
\name{season_name_short}
\alias{season_name_short}
\title{Short season name}
\format{
An object of class \code{character} of length 40.
}
\usage{
season_name_short
}
\description{
Short season name
}
\keyword{datasets}
