% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{season_palette}
\alias{season_palette}
\title{Gets season colour palette}
\usage{
season_palette(season_number)
}
\arguments{
\item{season_number}{Season number to extract colours for}
}
\value{
Returns a vector of hex codes
}
\description{
Helper function which simply returns the desired seasons tribe colours.
}
\examples{
\dontrun{
season_palette(40)
}
}
