% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{season_palettes}
\alias{season_palettes}
\title{Season palettes}
\format{
This nested data frame contains the following columns:
\describe{
\item{\code{version}}{Country code for the version of the show}
\item{\code{version_season}}{Version season key}
\item{\code{season_name}}{The season name}
\item{\code{season}}{The season number}
\item{\code{palette}}{The season palette}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
season_palettes
}
\description{
A dataset containing palettes generated from the season logos
}
\keyword{datasets}
