% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{season_participants}
\alias{season_participants}
\title{Season Participants}
\format{
A data frame with 39 rows and 4 variables:
\describe{
\item{\code{season}}{integer Season number}
\item{\code{participant_id}}{character The ID for the participant in the pool}
\item{\code{participant_full_name}}{character First and last name of the participant in the pool}
\item{\code{picking_order}}{double The order that the participant makes their pick}
}
}
\usage{
season_participants
}
\description{
Participants by season and their picking order
}
\details{
DETAILS
}
\keyword{datasets}
