% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{set_participant_picking_order}
\alias{set_participant_picking_order}
\title{Get Participant Picking Order}
\usage{
set_participant_picking_order(participants_enriched, season_picks)
}
\arguments{
\item{participants_enriched}{Participants data with full name joined in}

\item{season_picks}{Season picks dataset}
}
\description{
For each season, set the picking order. If no previous season, pick randomly.
If there's a new participant, they are placed last.
}
