% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_palette}
\alias{show_palette}
\title{Show a season logo palette}
\usage{
show_palette(version_season, n = NULL, type = "logo")
}
\arguments{
\item{version_season}{The version / season key e.g. `US42`}

\item{n}{The number of colours to view in the palette}

\item{type}{Either `logo` or `tribe`. Currently only for `logo`}
}
\value{
A ggplot2 graphic with the palette and logo
}
\description{
Easily view a palette for a given season and version including the log for reference
}
\examples{

show_palette("US43", n = 6)
}
