% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{still_alive}
\alias{still_alive}
\title{Still alive}
\usage{
still_alive(.vs, .n_boots = NULL, .ep = NULL, .at = "end")
}
\arguments{
\item{.vs}{Version season}

\item{.n_boots}{Number of boots}

\item{.ep}{Episode to evaluate who is alive.}

\item{.at}{Either 'start' or 'end'. If 'start' the flag will indicate who is
alive at the start of the episode. If 'end' it will indicate who is alive at
the end of the episode i.e. after tribal council.}
}
\value{
data frame
}
\description{
Finds the set of players that are still alive at either the start or end of
an episode, or given a set number of boots.
}
