% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{ui_create_picking_order}
\alias{ui_create_picking_order}
\title{Format Picking Order}
\usage{
ui_create_picking_order(szn = default_season())
}
\arguments{
\item{szn}{Season number. Defaults to \code{default_season()}}
}
\description{
Uses the results from the previous season to determine the current seasons
picking order. Collapses the participants names into a usable UI format
}
