% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{ui_create_picks_table}
\alias{ui_create_picks_table}
\title{Format Season Picks Table}
\usage{
ui_create_picks_table(szn = default_season(), picks_only = FALSE)
}
\arguments{
\item{szn}{Season number. Defaults to \code{default_season()}}

\item{picks_only}{Logical, whether or not to only show castaways that have
been picked by a contestant}
}
\description{
Formats data for the UI picks table
}
