% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{ui_create_season_pots_plot}
\alias{ui_create_season_pots_plot}
\title{Create Season Pots Plot}
\usage{
ui_create_season_pots_plot()
}
\description{
Creates season pots plot to show the pot for each season
}
