% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivor.R
\docType{data}
\name{webpages}
\alias{webpages}
\title{List of Wikipedia pages for each season}
\format{
An object of class \code{list} of length 40.
}
\usage{
webpages
}
\description{
List of Wikipedia pages for each season
}
\keyword{datasets}
