% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{which_versions}
\alias{which_versions}
\title{Checks which versions are available in the data}
\usage{
which_versions()
}
\value{
A message indicating which versions are loaded
}
\description{
By default on install only US versions are available however
using `import_non_us_data` it will import all other Non-US
versions available. Running this function will quickly tell
you which versions have been loaded.
}
\examples{
which_versions()
}
