% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{advantage_details}
\alias{advantage_details}
\title{Advantage Details}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{version}}{Country code for the version of the show}
  \item{\code{version_season}}{Version season key}
  \item{\code{season_name}}{The season name}
  \item{\code{season}}{The season number}
  \item{\code{advantage_id}}{The ID / primary key of the advantage}
  \item{\code{advantage_type}}{Advantage type e.g. hidden immunity idol, extra vote, steal a vote, etc}
  \item{\code{clue_details}}{Details if a clue existed for the advantage and if so where was the clue found}
  \item{\code{location_found}}{The location the idol or advantage was found}
  \item{\code{conditions}}{Extra details about the unique conditions of the idol or advantage}
  \item{\code{votes_nullified}}{In the case of hidden immunity idols this is the count of how many votes were nullified when played}
}
}
\usage{
advantage_details
}
\description{
A dataset containing the details and characteristics of each idol and advantage
}
\keyword{datasets}
