% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{auction_details}
\alias{auction_details}
\title{Survivor Auction Details}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{version}}{Country code for the version of the show}
  \item{\code{version_season}}{Version season key}
  \item{\code{season_name}}{The season name}
  \item{\code{season}}{The season number}
  \item{\code{item}}{Item number}
  \item{\code{item_description}}{Item description}
  \item{\code{castaway}}{Castaway}
  \item{\code{castaway_id}}{Castaway ID}
  \item{\code{covered}}{If the item was covered or not}
  \item{\code{cost}}{The amount paid for the item}
  \item{\code{money_remaining}}{How much money the castaway has remaining}
  \item{\code{auction_num}}{If the same item is auctioned for a second time it has a value of 2}
  \item{\code{participated}}{The names of castaways that could participate in the purchased item e.g. sharing a tub of peanut butter with the tribe}
  \item{\code{notes}}{Additional notes}
}
}
\usage{
auction_details
}
\description{
The details of the items purchased at the Survivor Auction
}
\keyword{datasets}
