% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{castaways}
\alias{castaways}
\title{Castaways}
\format{
Data frame:
\describe{
  \item{season}{The season the idol was found}
  \item{castaway}{Name of the castaway in possession of the idol. May not be the person who found it}
  \item{gender}{Gender of castaway}
  \item{finish_position}{Finishing position e.g. 1 is the winner}
  \item{days}{Number of days the castaway survived}
  \item{days_on_exile}{Number of days on Exile Island}
}
}
\source{
\url{tba}
}
\usage{
castaways
}
\description{
A dataset containing details on the castaways for each season.
}
\keyword{datasets}
