% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{castaways}
\alias{castaways}
\title{Castaways}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{version}}{Country code for the version of the show}
  \item{\code{version_season}}{Version season key}
  \item{\code{season}}{Sesaon number}
  \item{\code{season_name}}{Season name}
  \item{\code{full_name}}{Full name of the castaway}
  \item{\code{castaway_id}}{ID of the castaway (primary key). Consistent across seasons and name changes e.g. Amber Brkich / Amber Mariano. The first two letters reference the country of the version played e.g. US, AU (TBA).}
  \item{\code{castaway}}{Name of castaway. Generally this is the name they were most commonly referred to
  or nickname e.g. no one called Coach, Benjamin. He was simply Coach}
  \item{\code{age}}{Age of the castaway during the season they played}
  \item{\code{city}}{City of residence during the season they played}
  \item{\code{state}}{State of residence during the season they played}
  \item{\code{episode}}{Episode number}
  \item{\code{day}}{Number of days the castaway survived. A missing value indicates they later returned to the game that season}
  \item{\code{order}}{Boot order. Order in which castaway was voted out e.g. 5 is the 5th person voted of the island}
  \item{\code{result}}{Final result}
  \item{\code{result_number}}{Result number i.e. the final place. NA for castaways that were voted out but later returned e.g. Redemption Island}
  \item{\code{jury_status}}{Jury status}
  \item{\code{original_tribe}}{Original tribe name}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
castaways
}
\description{
A dataset containing details on the results for every castaway and season
}
\details{
If the original \code{castaway_id} is desired simply extract the digits from the ID e.g.
\code{castaway_id = as.numeric(str_extract(castaway_id, '[:digit:]+'))} in a mutate step.
}
\examples{
library(dplyr)
castaways \%>\%
  filter(season == 40)
}
\keyword{datasets}
