% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{castaways}
\alias{castaways}
\title{Castaways}
\format{
Data frame:
\describe{
  \item{season}{Sesaon number}
  \item{season_name}{Season name}
  \item{full_name}{Full name of the castaway}
  \item{castaway}{Name of castaway}
  \item{gender}{Gender of castaway}
  \item{age}{Age of the castaway}
  \item{city}{City of residence during the season they played}
  \item{state}{State of residence during the season they played}
  \item{day}{Number of days the castaway survived. A missing value indicates they later returned to the game that season}
  \item{order}{Order in which castaway was voted out e.g. 5 is the 5th person voted of the island}
  \item{result}{Final result}
  \item{jury_status}{Jury status}
  \item{original_tribe}{Original tribe name}
  \item{merged_tribe}{Merged tribe name}
  \item{swapped_tribe}{Swapped tribe name}
  \item{swapped_tribe2}{second swapped tribe in the event of a second tribe swap or other tribe restructure such as absorbed tribe, outcasts, etc}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
castaways
}
\description{
A dataset containing details on the castaways for each season.
}
\examples{
\dontrun{
castaways \%>\%
  filter(season == 40)
}
}
\keyword{datasets}
