% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{challenges}
\alias{challenges}
\title{Challenges}
\format{
This nested data frame contains the following columns:
\describe{
  \item{\code{season_name}}{The season name}
  \item{\code{season}}{The season number}
  \item{\code{episode}}{Episode number}
  \item{\code{title}}{Episode title}
  \item{\code{day}}{The day of the tribal council}
  \item{\code{challenge_type}}{The challenge type e.g. immunity, reward, etc}
  \item{\code{winners}}{The list of winners. Either the list of people in the tribe which won, list of people that participated on the reward or the individual winner}
  \item{\code{winning_tribe}}{Nmae of the winner tribe. \code{NA} during the merge}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
challenges
}
\description{
A dataset detailing the challenges played including reward and immunity challenges.
}
\keyword{datasets}
