% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_jury_votes}
\alias{clean_jury_votes}
\title{Clean jury vote data}
\usage{
clean_jury_votes(df, season, season_name)
}
\arguments{
\item{df}{Data frame}

\item{season}{Season number}

\item{season_name}{Season name}
}
\value{

}
\description{
Clean jury vote data
}
\examples{
\dontrun{}
}
