% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{confessionals}
\alias{confessionals}
\title{Confessionals}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{season_name}}{The season name}
  \item{\code{season}}{The season number}
  \item{\code{episode}}{Episode number}
  \item{\code{castaway}}{Name of the castaway}
  \item{\code{castaway_id}}{ID of the castaway (primary key). Consistent across seasons and name changes e.g. Amber Brkich / Amber Mariano}
  \item{\code{confessional_count}}{The count of confessionals for the castaway during the episode}
}
}
\usage{
confessionals
}
\description{
A dataset containing the count of confessionals per castaway per episode. A confessional is when
the castaway is speaking directly to the camera about their game.
}
\keyword{datasets}
