% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{create_season_picks}
\alias{create_season_picks}
\title{Create Season Picks Data Frame}
\usage{
create_season_picks(all_data = NULL, augment_with_historical = TRUE)
}
\arguments{
\item{all_data}{Named list of tabs and their data from the survivor googlesheet}

\item{augment_with_historical}{Logical, whether or not to add historical data.
See \code{?historical_data} for more info}
}
\description{
Main method for creating the season picks table with all transformations applied
}
