% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{create_season_picks}
\alias{create_season_picks}
\title{Create Season Picks Data Frame}
\usage{
create_season_picks(
  using = c("googlesheets", "configs"),
  config_path = "data-raw/configs",
  cost_per_day = 1,
  sole_survivor_bonus = 5
)
}
\arguments{
\item{using}{Either googlesheets or configs}

\item{config_path}{Path to the config files, defaults to the right path}

\item{cost_per_day}{Integer, number of dollars owed per day using the difference
in days compared to the person who's castaway lasted the longest}

\item{sole_survivor_bonus}{Integer, number of dollars to add to each persons
payment if the winners pick was the sole survivor}
}
\description{
Main method for creating the season picks table with all transformations applied
}
