% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{.get_season_payments}
\alias{.get_season_payments}
\title{Get Participant Payments}
\usage{
.get_season_payments(picks)
}
\arguments{
\item{picks}{Picks dataset}
}
\description{
Get Participant Payments
}
