% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{.process_seasons__googlesheets}
\alias{.process_seasons__googlesheets}
\title{Process All Seasons Using Google Sheets}
\usage{
.process_seasons__googlesheets()
}
\description{
Process the data in the survivor Google Sheet
}
