% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_finalist}
\alias{filter_finalist}
\title{Filter to finalists}
\usage{
filter_finalist(df)
}
\arguments{
\item{df}{Data frame. Requires \code{version_season} and \code{castaway_id}.}
}
\value{
A data frame filtered to the finalists
}
\description{
Filters a data set to the finalists of a given season.
}
\examples{

library(survivoR)
library(dplyr)

confessionals |>
  filter_finalist()

}
