% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{format_picks_table}
\alias{format_picks_table}
\title{Format Season Picks Table}
\usage{
format_picks_table(szn = default_season())
}
\arguments{
\item{szn}{Season number. Defaults to \code{default_season()}}
}
\description{
Selects, renames, and formats relevant columns for the app
}
