% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_confessional_timing}
\alias{get_confessional_timing}
\title{Confessional time}
\usage{
get_confessional_timing(path, .vs, .episode, folder = "Final")
}
\arguments{
\item{path}{Parent director where the files are stored. This parth should include 3 sub-directories
Final, Staging and Notes.}

\item{.vs}{Version season}

\item{.episode}{Episode}

\item{folder}{The folder to read the text file. Default is 'Final'}
}
\value{
data frame
}
\description{
Confessional time
}
\examples{
# After running app and recording confessionals, run...
# Example from a saved timing file

# get_confessional_timing(path = , .vs = "US44", .episode = 13)
}
