% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_confessional_timing}
\alias{get_confessional_timing}
\title{Confessional time}
\usage{
get_confessional_timing(paths, .vs, .episode, .mda = 3)
}
\arguments{
\item{paths}{Paths to the csv file containing all the time stamps from the Shiny app}

\item{.vs}{Version season}

\item{.episode}{Episode}

\item{.mda}{Missing duration adjustment in seconds. If either start or stop is missing from the
records, the missing value is imputed with a 3 second adjustment by default.}
}
\value{
data frame
}
\description{
Takes the output of the times recorded from the Shiny app and aggregates to the final
confessional times and confessional counts. \code{confessional_time} is the total duration
in seconds for the episode. \code{confessional_count} is the number of confessionals
recorded to be at least 10 seconds apart.
}
\examples{
# After running app and recording confessionals, run...
# Example from a saved timing file

path <- system.file(package = "survivoR", "confessionalDash/US4412.csv")
get_confessional_timing(path = path, .vs = "US44", .episode = 12)
}
