% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_confessional_timing}
\alias{get_confessional_timing}
\title{Confessional time}
\usage{
get_confessional_timing(paths, .vs, .episode)
}
\arguments{
\item{.vs}{Version season}

\item{.episode}{Episode}

\item{path}{Parent director where the files are stored. This parth should include 3 sub-directories
Final, Staging and Notes.}
}
\value{
data frame
}
\description{
Confessional time
}
\examples{
# After running app and recording confessionals, run...
# Example from a saved timing file

path <- system.file(package = "survivoR", "confessionalDash/2023-05-18 19h26m53s US4412.csv")
get_confessional_timing(path = path, .vs = "US44", .episode = 12)
}
