% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{gs_auth}
\alias{gs_auth}
\title{Authenticate With Google Sheets}
\usage{
gs_auth(cache = ".secrets", email = "elampsart@gmail.com", deauth_mode = FALSE)
}
\arguments{
\item{cache}{Where the secret is stored}

\item{email}{Email to authenticate with}

\item{deauth_mode}{Whether or not to run only with \code{\link[googlesheets4:gs4_deauth]{googlesheets4::gs4_deauth()}}}
}
\description{
Authenticate With Google Sheets
}
