% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hidden_idols}
\alias{hidden_idols}
\title{Hidden immunity idols}
\format{
A dataset containing the details of who found a hidden immunity idol,
when and when or if they played it:
\describe{
  \item{season}{The season the idol was found}
  \item{castaway}{Name of the castaway in possession of the idol. May not be the person who found it}
  \item{idol_number}{}
  \item{idols_held}{}
}
}
\source{
\url{tba}
}
\usage{
hidden_idols
}
\description{
A dataset containing the complete history of hidden immunity idols
}
\keyword{datasets}
