% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hidden_idols}
\alias{hidden_idols}
\title{(Deprecated) Hidden Immunity Idols}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{season_name}}{The season name}
  \item{\code{season}}{The season number}
  \item{\code{castaway_id}}{ID of the castaway (primary key). Consistent across seasons and name changes e.g. Amber Brkich / Amber Mariano. The first two letters reference the country of the version played e.g. US, AU.}
  \item{\code{castaway}}{Name of the castaway}
  \item{\code{idol_number}}{Indicates whether it is the first, second, etc idol found in the season}
  \item{\code{idols_held}}{The number of idols held by the castaway}
  \item{\code{votes_nullified}}{The number of votes nullified by the idol}
  \item{\code{day_found}}{The day the idol was found}
  \item{\code{day_played}}{The day of the tribal council}
  \item{\code{legacy_advantage}}{If the idol was a legacy advantage or not}
}
}
\source{
\url{https://survivor.fandom.com/wiki/Hidden_Immunity_Idol}
}
\usage{
hidden_idols
}
\description{
This data set has be replaced with the advantage data sets which contain more details
and easier to use structure
}
\keyword{datasets}
