% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{immunity}
\alias{immunity}
\title{Immunity challenges}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{episode}{Episode number of the immunity challenge was played}
  \item{title}{Episode title}
  \item{immunity}{Winners of the immunity challenge}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
immunity
}
\description{
A dataset containing details on the immunity challenges for each season
}
\details{
Contains details on tribal immunity and individual immunity. Not hidden immunity however.
}
\examples{
\dontrun{
immunity
immunity \%>\%
  unnest(c(immunity))
}
}
\keyword{datasets}
