% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jury_votes}
\alias{jury_votes}
\title{Jury votes}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{castaway}{Name of the castaway}
  \item{finalist}{The finalists for which a vote can be placed}
  \item{vote}{Vote}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
jury_votes
}
\description{
A dataset containing details on the final jury votes to determine the winner for each season
}
\examples{
\dontrun{
jury_votes \%>\%
  filter(season == 40) \%>\%
  group_by(finalist) \%>\%
  summarise(votes = sum(vote))
  }
}
\keyword{datasets}
