% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{launch_confessional_app}
\alias{launch_confessional_app}
\title{Launch Confessional App}
\usage{
launch_confessional_app(path, browser = TRUE)
}
\arguments{
\item{path}{Parent director for where the files will be saved.}

\item{browser}{Open in browser instead of viewer. Default \code{TRUE}}
}
\value{
Shiny app
}
\description{
Opens a viewer or browser to time castaway confessionals
}
\examples{

# Run app

# launch_confessional_app()

}
