% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{launch_confessional_app}
\alias{launch_confessional_app}
\title{Launch Confessional App}
\usage{
launch_confessional_app(browser = TRUE)
}
\arguments{
\item{browser}{Open in browser instead of viewer. Default \code{TRUE}}
}
\value{
Shiny app
}
\description{
Launches the confessional timing app in either a browser or viewer. Default is set to
browser. The user is required to provide a path for which the time stanps are recoreded.
}
\examples{
## Only run this example in interactive R sessions

if(interactive()) {
  # launch app
  launch_confessional_app()
}

}
