% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons.R
\name{make_season_wiki_link}
\alias{make_season_wiki_link}
\title{Make Season Wiki Link}
\usage{
make_season_wiki_link(szn = default_season())
}
\arguments{
\item{szn}{Season number}
}
\description{
Make Season Wiki Link
}
