% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{process_season_config}
\alias{process_season_config}
\title{Process Season Config}
\usage{
process_season_config(config, participants)
}
\arguments{
\item{config}{Season config file parsed using \code{yaml::read_yaml}}

\item{participants}{Config file of participants in the pool parsed using \code{yaml::read_yaml}}
}
\description{
Process season config into a data frame
}
