% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_init.R
\name{process_seasons}
\alias{process_seasons}
\title{Process Seasons Table}
\usage{
process_seasons(participants, castaways, picks, seasons)
}
\arguments{
\item{participants}{Participants data from googlesheet}

\item{castaways}{Castaway data from googlesheet}

\item{picks}{Picks data from googlesheet}

\item{seasons}{Seasons data from googlesheet}
}
\description{
Creates the base season picks table using either the data in googlesheets
or the config files
}
