% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_data.R
\name{process_seasons}
\alias{process_seasons}
\title{Process Seasons Table}
\usage{
process_seasons(
  using = c("googlesheets", "configs"),
  config_path = "data-raw/configs"
)
}
\arguments{
\item{using}{Either googlesheets or configs}

\item{config_path}{Path to the config files}
}
\description{
Creates the base season picks table using either the data in googlesheets
or the config files
}
