% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rewards}
\alias{rewards}
\title{Reward challenges}
\format{
Nested tidy data frame:
\describe{
  \item{season_name}{Season name}
  \item{season}{Sesaon number}
  \item{episode}{Episode number of when the reward challenge was played}
  \item{title}{Episode title}
  \item{Reward}{Winners of the reward challenge. Tidy data frame. The list of castaway include all those that participated in the reward rather than simply the castaway that won the challenge.}
}
}
\source{
\url{tba}
}
\usage{
rewards
}
\description{
A dataset containing details on the reward challenges for each season
}
\details{
This is a nested data frame since more than one person can win the reward
}
\examples{
\dontrun{
rewards
rewards \%>\%
  unnest(c(reward))
}
}
\keyword{datasets}
