% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{survivor_pal}
\alias{survivor_pal}
\alias{scale_fill_survivor}
\alias{scale_colour_survivor}
\title{Survivor palette}
\usage{
survivor_pal(season, scale_type = "d", reverse = FALSE, tribe = NULL, ...)

scale_fill_survivor(season, scale_type = "d", reverse = FALSE, ...)

scale_colour_survivor(season, scale_type = "d", reverse = FALSE, ...)
}
\arguments{
\item{season}{Season number}

\item{scale_type}{Discrete or continuous. Input 'd' / 'c'.}

\item{reverse}{Logical. Reverse the palette?}

\item{...}{Dots}
}
\value{
Scale functions for ggplot2




}
\description{
To create scale functions for ggplot. Given a season of Survivor, a palette
is created from the tribe colours for that season including the merged tribe.
}
\examples{
\dontrun{
season_summary \%>\%
  select(season, viewers_premier, viewers_finale, viewers_reunion, viewers_mean) \%>\%
  pivot_longer(cols = -season, names_to = "episode", values_to = "viewers") \%>\%
  mutate(
    episode = to_title_case(str_replace(episode, "viewers_", ""))
  ) \%>\%
  ggplot(aes(x = season, y = viewers, colour = episode)) +
  geom_line() +
  geom_point(size = 2) +
  theme_minimal() +
  scale_colour_survivor(16)
 }
}
