% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{season_picks}
\alias{season_picks}
\title{Season Picks}
\format{
A data frame with 36 rows and 8 variables:
\describe{
\item{\code{season}}{integer Season number}
\item{\code{person_id}}{character The ID for the person in the pool}
\item{\code{person_first}}{character First name of the person in the pool}
\item{\code{person_last}}{character Last name of the person in the pool}
\item{\code{castaway_id}}{character ID for the castaway}
\item{\code{castaway_name}}{character Full name of the castaway}
\item{\code{castaway_finish_day}}{integer The day the castaway was voted out}
\item{\code{castaway_finish_inverse_placement}}{integer The inverse placement of the castaway}
\item{\code{eliminated}}{logical Whether or not survivor has been eliminated}
\item{\code{castaway_finish_placement}}{double Rank based on when castaway was voted out}
\item{\code{sole_survivor}}{logical Whether or not the castaway was the sole survivor}
}
}
\usage{
season_picks
}
\description{
Picks across all seasons recorded in this repo
}
\details{
Season configs processed into a table
}
\keyword{datasets}
