% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{season_summary}
\alias{season_summary}
\title{Season summary}
\format{
Data frame:
\describe{
  \item{season_name}{Season name}
  \item{season}{Sesaon number}
  \item{location}{Location of the season}
  \item{country}{Country the season was held}
  \item{tribe_setup}{Initial setup of the tribe e.g. heroes vs Healers vs Hustlers}
  \item{full_name}{Winner of the season's full name}
  \item{winner}{Winner of the season}
  \item{runner_ups}{runner ups for the season. Nested data from given there may be 2 runner ups and this preserves the grain of the data being a season}
  \item{final_vote}{Final vote allocation. See the \code{jury_votes} dataset for better aggregation of this data}
  \item{timeslot}{Timeslot of the show in the US}
  \item{premiered}{Date the first episode aired}
  \item{ended}{Date the season ended}
  \item{viewers_premier}{Number of viewers (millions) who tuned in for the premier}
  \item{viewers_finale}{Number of viewers (millions) who tuned in for the finale}
  \item{viewers_reunion}{Number of viewers (millions) who tuned in for the reunion}
  \item{viewers_mean}{Average number of viewers (millions) who tuned in over the season}
  \item{rank}{Season rank}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
season_summary
}
\description{
A dataset containing a summary of all 40 seasons of survivor
}
\examples{
\dontrun{
View(season_summary)
}
}
\keyword{datasets}
