% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tribe_mapping}
\alias{tribe_mapping}
\title{Tribe mapping}
\format{
This data frame contains the following columns:
\describe{
  \item{\code{season_name}}{The season name}
  \item{\code{season}}{The season number}
  \item{\code{day}}{The day of the tribal council}
  \item{\code{castaway}}{Castaway}
  \item{\code{tribe}}{Name of the tribe the castaway was on}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
tribe_mapping
}
\description{
[dev] A mapping for castaways to tribes for each day (day being the day of the tribal council)
This is useful for observing who is on what tribe throughout the game.
}
\keyword{datasets}
