% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{viewers}
\alias{viewers}
\title{Viewer data}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{episode}{Episode number of the reward challenge}
  \item{title}{Episode title}
  \item{immunity}{Winners of the immunity challenge}
}
}
\source{
\url{tba}
}
\usage{
viewers
}
\description{
A dataset containing details on the viewer history for each eason
}
\keyword{datasets}
