% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{viewers}
\alias{viewers}
\title{Viewers}
\format{
data frame
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season the idol was found}
  \item{episode_number_overall}{The cumulative episode number}
  \item{episode}{Episode number for the season}
  \item{title}{Episode title}
  \item{episode_date}{Date the episode aired}
  \item{viewers}{Number of viewers (millions) who tuned in.}
  \item{rating_18_49}{TV rating for the 18-49 aged group}
  \item{share_18_49}{TV share for the 18_49 aged group}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
viewers
}
\description{
A dataset containing the viewer history for each season
}
\keyword{datasets}
