% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{viewers}
\alias{viewers}
\title{Viewers}
\format{
data frame
\describe{
  \item{\code{season_name}}{The season name}
  \item{\code{season}}{The season the idol was found}
  \item{\code{episode_number_overall}}{The cumulative episode number}
  \item{\code{episode}}{Episode number for the season}
  \item{\code{title}}{Episode title}
  \item{\code{episode_date}}{Date the episode aired}
  \item{\code{viewers}}{Number of viewers (millions) who tuned in}
  \item{\code{rating_18_49}}{TV rating for the 18-49 aged group}
  \item{\code{share_18_49}}{TV share for the 18-49 aged group}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
viewers
}
\description{
A dataset containing the viewer history for each season and episode
}
\keyword{datasets}
