% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vote_history}
\alias{vote_history}
\title{Vote history}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{episode}{Episode number of the reward challenge}
  \item{day}{Number of days the castaway survived. A missing value indicates they later returned to the game that season}
  \item{castaway}{Name of the castaway}
  \item{tribe_status}{The status of the tribe e.g. original tribe, swapped tribe, merged tribe, etc}
  \item{vote}{The castaway for which the vote was cast}
  \item{voted_out}{Who was voted out}
  \item{order}{Finishing position e.g. 1 is the winner}
}
}
\source{
\url{tba}
}
\usage{
vote_history
}
\description{
A dataset containing details on the vote history for each season
}
\keyword{datasets}
