% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vote_history}
\alias{vote_history}
\title{Vote history}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{episode}{Episode number of the reward challenge}
  \item{title}{Episode title}
  \item{immunity}{Winners of the immunity challenge}
}
}
\source{
\url{tba}
}
\usage{
vote_history
}
\description{
A dataset containing details on the vote history for each season
}
\keyword{datasets}
