% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vote_history}
\alias{vote_history}
\title{Vote history}
\format{
Data frame:
\describe{
  \item{season_name}{The season_name}
  \item{season}{The season number}
  \item{episode}{Episode number of the reward challenge}
  \item{day}{Number of days the castaway survived. A missing value indicates they later returned to the game that season}
  \item{tribe_status}{The status of the tribe e.g. original tribe, swapped tribe, merged tribe, etc. See details for more}
  \item{castaway}{Name of the castaway}
  \item{immunity}{Type of immunity held by the castaway at the time of the vote e.g. individual, hidden}
  \item{vote}{The castaway for which the vote was cast}
  \item{nullified}{Was the vote nullified by a hidden immunity idol? Logical.}
  \item{voted_out}{Who was voted out}
  \item{order}{The order in which the castaway was voted out}
  \item{vote_order}{The the case of ties this indicates the order of the votes}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Survivor_(American_TV_series)}
}
\usage{
vote_history
}
\description{
A dataset containing details on the vote history for each season
}
\details{
This data frame contains a complete history of votes cast across all seasons of Survivor. While there are consistent
events across the seasons there are some unique events such as the 'mutiny' in Survivor: Cook Islands (season 13)
or the 'Outcasts' in Survivor: Pearl Islands (season 7). For maintaining a standard, whenever there has been a change
in tribe for the castaways it has been recorded as \code{tribe_status == 'swapped'}. Subsequent changes are recorded with
a digit. This includes absorbed tribes e.g. Stephanie was 'absorbed' in Survivor: Palau (season 10) and when 3 tribes are
reduced to 2. These cases are still considered 'swapped' to indicate a change in tribe status. 'Swapped' is used as the
term since 'the tribe swap' is a typical recurring milestone in each season of Survivor.

Some events result in a castaway attending tribal but not voting. These are recorded as
\describe{
  \item{Win}{The castaway won the fire challenge}
  \item{Lose}{The castaway lose the fire challenge}
  \item{None}{The castaway did not cast a vote. This may be due to a vote steal or some other means}
  \item{Immune}{The castaway did not vote but were immune from the vote}
}

Where a castaway has \code{immunity == 'hidden'} means that player is protected by a hidden immunity idol. It may not
necessarily mean they played the idol, the idol may have been played for them. While the nullified votes data is complete
the \code{immunity} data does not include those who had immunity but did not receive a vote. This is a TODO
}
\examples{
\dontrun{
# The number of times Tony voted for each castaway in Survivor: Winners at War
vote_history \%>\%
  filter(
    season == 40,
    castaway == "Tony"
  ) \%>\%
  count(vote)
}
}
\keyword{datasets}
